python/requirements.txt
Flask==2.3.3
requests==2.31.0
PyJWT==2.8.0
cryptography==41.0.3
python-jose==3.3.0
oauthlib==3.2.2
requests-oauthlib==1.3.1
requests-ntlm==1.2.0
requests-hawk==1.0.0
requests-aws4auth==1.2.3
akamai-edgegrid==1.1.1 